LPD-58417 Merge FlushIndexRequestExecutor/FlushIndexRequestExecutorImpl into pojos
