--- modules/apps/portal-search-solr8/portal-search-solr8-impl/src/main/java/com/liferay/portal/search/solr8/internal/search/engine/adapter/index/FlushIndexRequestExecutorImpl.java
@@ -1,25 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.solr8.internal.search.engine.adapter.index;
-
-import com.liferay.portal.search.engine.adapter.index.FlushIndexRequest;
-import com.liferay.portal.search.engine.adapter.index.FlushIndexResponse;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Bryan Engler
- */
-@Component(service = FlushIndexRequestExecutor.class)
-public class FlushIndexRequestExecutorImpl
-	implements FlushIndexRequestExecutor {
-
-	@Override
-	public FlushIndexResponse execute(FlushIndexRequest flushIndexRequest) {
-		throw new UnsupportedOperationException();
-	}
-
-}
